Updated ProductPharmacy seeder to work with server

+++ database/seeds/ProductPharmacySeeder.php
@@ -14,36 +14,34 @@ class ProductPharmacySeeder extends Seeder
      * @return void
      */
 
-    public function key_value_pair_exists(array $haystack, $key, $value) {
-        return array_key_exists($key, $haystack) &&
-               $haystack[$key] == $value;
-    }
-
     public function run()
     {
         //  First, check the products_pharmacies table for content
         if (DB::table('products_pharmacies')->get()->count() != 0) {
 
-            //  Reset the id counting and clears the table
-            //  Remove and re add the foreign key checks to clear the table id incerment
-            DB::statement('SET FOREIGN_KEY_CHECKS=0;');
-            DB::table('products_pharmacies')->truncate();
-            DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+            // Remove and re add the foreign key checks to clear the table id incerment
+            if (DB::connection()->getDatabaseName() == 'curey_db'){
+                DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+                // Reset the id counting and clears the table
+                DB::table('products_pharmacies')->truncate();
+                DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+            }
+            else {
+                DB::statement('TRUNCATE products_pharmacies RESTART IDENTITY CASCADE;');
+            }
         }
 
-        //  Get IDs available in database
-        $products = Product::get()->toArray();
-        $pharmacies = Pharmacy::get()->toArray();
-
         //  Remove the foreign key checks to change the IDs
-        DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+        if (DB::connection()->getDatabaseName() == 'curey_db'){
+            DB::statement('SET FOREIGN_KEY_CHECKS=0;');
+        }
 
         //  Change the i for numbers of rows generated
         for ($i = 0; $i < 100; $i++) {
             
             //  Assign random IDs in every iteration
-            $product = $products[array_rand($products)]['id'];
-            $pharmacy = $pharmacies[array_rand($pharmacies)]['id'];
+            $product = DB::table('products')->inRandomOrder()->first() -> id;
+            $pharmacy = DB::table('pharmacies')->inRandomOrder()->first() -> id;
 
             //  Check if the product ID exists in the database, if not, adds the record
             if (ProductPharmacy::where('product_id', $product)->count() == 0){
@@ -70,7 +68,9 @@ class ProductPharmacySeeder extends Seeder
         }
 
         // Re add the foreign key checks
-        DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+        if (DB::connection()->getDatabaseName() == 'curey_db'){
+            DB::statement('SET FOREIGN_KEY_CHECKS=1;');
+        }
 
         $this->command->info('    ---------------------------------------');
         $this->command->info('  Products_Pharmacies table updated ¯\_(ツ)_/¯');
